+++ sdk/monitor/azure-monitor-query/azure/monitor/query/_generated/operations/_operations.py
@@ -1,4 +1,4 @@
-# pylint: disable=too-many-lines
+# pylint: disable=too-many-lines,too-many-statements
 # coding=utf-8
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
@@ -271,6 +271,7 @@ class QueryOperations:
     def get(
         self, workspace_id: str, *, query: str, timespan: Optional[datetime.timedelta] = None, **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query.
 
         Executes an Analytics query for data.
@@ -395,6 +396,7 @@ class QueryOperations:
         content_type: str = "application/json",
         **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query.
 
         Executes an Analytics query for data. `Here
@@ -485,12 +487,13 @@ class QueryOperations:
     def execute(
         self,
         workspace_id: str,
-        body: IO,
+        body: IO[bytes],
         *,
         prefer: Optional[str] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query.
 
         Executes an Analytics query for data. `Here
@@ -503,7 +506,7 @@ class QueryOperations:
         :param body: The Analytics query. Learn more about the `Analytics query syntax
          <https://azure.microsoft.com/documentation/articles/app-insights-analytics-reference/>`_.
          Required.
-        :type body: IO
+        :type body: IO[bytes]
         :keyword prefer: Optional. The prefer header to set server timeout, query statistics and
          visualization information. Default value is None.
         :paramtype prefer: str
@@ -566,7 +569,10 @@ class QueryOperations:
         """
 
     @distributed_trace
-    def execute(self, workspace_id: str, body: Union[JSON, IO], *, prefer: Optional[str] = None, **kwargs: Any) -> JSON:
+    def execute(
+        self, workspace_id: str, body: Union[JSON, IO[bytes]], *, prefer: Optional[str] = None, **kwargs: Any
+    ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query.
 
         Executes an Analytics query for data. `Here
@@ -578,8 +584,8 @@ class QueryOperations:
         :type workspace_id: str
         :param body: The Analytics query. Learn more about the `Analytics query syntax
          <https://azure.microsoft.com/documentation/articles/app-insights-analytics-reference/>`_. Is
-         either a JSON type or a IO type. Required.
-        :type body: JSON or IO
+         either a JSON type or a IO[bytes] type. Required.
+        :type body: JSON or IO[bytes]
         :keyword prefer: Optional. The prefer header to set server timeout, query statistics and
          visualization information. Default value is None.
         :paramtype prefer: str
@@ -712,6 +718,7 @@ class QueryOperations:
     def resource_get(
         self, resource_id: str, *, query: str, timespan: Optional[datetime.timedelta] = None, **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query using resource URI.
 
         Executes an Analytics query for data in the context of a resource. `Here
@@ -837,6 +844,7 @@ class QueryOperations:
         content_type: str = "application/json",
         **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query using resource ID.
 
         Executes an Analytics query for data in the context of a resource. `Here
@@ -926,12 +934,13 @@ class QueryOperations:
     def resource_execute(
         self,
         resource_id: str,
-        body: IO,
+        body: IO[bytes],
         *,
         prefer: Optional[str] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query using resource ID.
 
         Executes an Analytics query for data in the context of a resource. `Here
@@ -943,7 +952,7 @@ class QueryOperations:
         :param body: The Analytics query. Learn more about the `Analytics query syntax
          <https://azure.microsoft.com/documentation/articles/app-insights-analytics-reference/>`_.
          Required.
-        :type body: IO
+        :type body: IO[bytes]
         :keyword prefer: Optional. The prefer header to set server timeout, query statistics and
          visualization information. Default value is None.
         :paramtype prefer: str
@@ -1007,8 +1016,9 @@ class QueryOperations:
 
     @distributed_trace
     def resource_execute(
-        self, resource_id: str, body: Union[JSON, IO], *, prefer: Optional[str] = None, **kwargs: Any
+        self, resource_id: str, body: Union[JSON, IO[bytes]], *, prefer: Optional[str] = None, **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query using resource ID.
 
         Executes an Analytics query for data in the context of a resource. `Here
@@ -1019,8 +1029,8 @@ class QueryOperations:
         :type resource_id: str
         :param body: The Analytics query. Learn more about the `Analytics query syntax
          <https://azure.microsoft.com/documentation/articles/app-insights-analytics-reference/>`_. Is
-         either a JSON type or a IO type. Required.
-        :type body: JSON or IO
+         either a JSON type or a IO[bytes] type. Required.
+        :type body: JSON or IO[bytes]
         :keyword prefer: Optional. The prefer header to set server timeout, query statistics and
          visualization information. Default value is None.
         :paramtype prefer: str
@@ -1151,6 +1161,7 @@ class QueryOperations:
 
     @overload
     def batch(self, body: JSON, *, content_type: str = "application/json", **kwargs: Any) -> JSON:
+        # pylint: disable=line-too-long
         """Execute a batch of Analytics queries.
 
         Executes a batch of Analytics queries for data. `Here
@@ -1274,7 +1285,8 @@ class QueryOperations:
         """
 
     @overload
-    def batch(self, body: IO, *, content_type: str = "application/json", **kwargs: Any) -> JSON:
+    def batch(self, body: IO[bytes], *, content_type: str = "application/json", **kwargs: Any) -> JSON:
+        # pylint: disable=line-too-long
         """Execute a batch of Analytics queries.
 
         Executes a batch of Analytics queries for data. `Here
@@ -1282,7 +1294,7 @@ class QueryOperations:
         using POST with an Analytics query.
 
         :param body: The batch request body. Required.
-        :type body: IO
+        :type body: IO[bytes]
         :keyword content_type: Body Parameter content-type. Content type parameter for binary body.
          Default value is "application/json".
         :paramtype content_type: str
@@ -1365,15 +1377,16 @@ class QueryOperations:
         """
 
     @distributed_trace
-    def batch(self, body: Union[JSON, IO], **kwargs: Any) -> JSON:
+    def batch(self, body: Union[JSON, IO[bytes]], **kwargs: Any) -> JSON:
+        # pylint: disable=line-too-long
         """Execute a batch of Analytics queries.
 
         Executes a batch of Analytics queries for data. `Here
         <https://learn.microsoft.com/azure/azure-monitor/logs/api/batch-queries>`_ is an example for
         using POST with an Analytics query.
 
-        :param body: The batch request body. Is either a JSON type or a IO type. Required.
-        :type body: JSON or IO
+        :param body: The batch request body. Is either a JSON type or a IO[bytes] type. Required.
+        :type body: JSON or IO[bytes]
         :keyword content_type: Body Parameter content-type. Known values are: 'application/json'.
          Default value is None.
         :paramtype content_type: str
@@ -1545,6 +1558,7 @@ class QueryOperations:
     def resource_get_xms(
         self, resource_id: str, *, query: str, timespan: Optional[datetime.timedelta] = None, **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query using resource URI.
 
         Executes an Analytics query for data in the context of a resource. `Here
@@ -1670,6 +1684,7 @@ class QueryOperations:
         content_type: str = "application/json",
         **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query using resource ID.
 
         Executes an Analytics query for data in the context of a resource. `Here
@@ -1759,12 +1774,13 @@ class QueryOperations:
     def resource_execute_xms(
         self,
         resource_id: str,
-        body: IO,
+        body: IO[bytes],
         *,
         prefer: Optional[str] = None,
         content_type: str = "application/json",
         **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query using resource ID.
 
         Executes an Analytics query for data in the context of a resource. `Here
@@ -1776,7 +1792,7 @@ class QueryOperations:
         :param body: The Analytics query. Learn more about the `Analytics query syntax
          <https://azure.microsoft.com/documentation/articles/app-insights-analytics-reference/>`_.
          Required.
-        :type body: IO
+        :type body: IO[bytes]
         :keyword prefer: Optional. The prefer header to set server timeout, query statistics and
          visualization information. Default value is None.
         :paramtype prefer: str
@@ -1840,8 +1856,9 @@ class QueryOperations:
 
     @distributed_trace
     def resource_execute_xms(
-        self, resource_id: str, body: Union[JSON, IO], *, prefer: Optional[str] = None, **kwargs: Any
+        self, resource_id: str, body: Union[JSON, IO[bytes]], *, prefer: Optional[str] = None, **kwargs: Any
     ) -> JSON:
+        # pylint: disable=line-too-long
         """Execute an Analytics query using resource ID.
 
         Executes an Analytics query for data in the context of a resource. `Here
@@ -1852,8 +1869,8 @@ class QueryOperations:
         :type resource_id: str
         :param body: The Analytics query. Learn more about the `Analytics query syntax
          <https://azure.microsoft.com/documentation/articles/app-insights-analytics-reference/>`_. Is
-         either a JSON type or a IO type. Required.
-        :type body: JSON or IO
+         either a JSON type or a IO[bytes] type. Required.
+        :type body: JSON or IO[bytes]
         :keyword prefer: Optional. The prefer header to set server timeout, query statistics and
          visualization information. Default value is None.
         :paramtype prefer: str
@@ -2002,6 +2019,7 @@ class MetadataOperations:
 
     @distributed_trace
     def get(self, workspace_id: str, **kwargs: Any) -> JSON:
+        # pylint: disable=line-too-long
         """Gets metadata information.
 
         Retrieve the metadata information for the workspace, including its schema, functions, workspace
@@ -2403,6 +2421,7 @@ class MetadataOperations:
 
     @distributed_trace
     def post(self, workspace_id: str, **kwargs: Any) -> JSON:
+        # pylint: disable=line-too-long
         """Gets metadata information.
 
         Retrieve the metadata information for the workspace, including its schema, functions, workspace